Client package renaming
