+++ pom.xml
@@ -58,7 +58,7 @@
             <!-- Default configuration for running with: mvn clean javafx:run -->
             <id>default-cli</id>
             <configuration>
-              <mainClass>com.example.bookrecommenderdev/com.example.bookrecommenderdev.HelloApplication</mainClass>
+              <mainClass>com.example.bookrecommenderdev/com.example.bookrecommenderdev.client.HelloApplication</mainClass>
               <launcher>app</launcher>
               <jlinkZipName>app</jlinkZipName>
               <jlinkImageName>app</jlinkImageName>
