+++ src/main/java/com/example/bookrecommenderdev/client/model/EntryValutazione.java
@@ -1,4 +1,4 @@
-package com.example.bookrecommenderdev.model;
+package com.example.bookrecommenderdev.client.model;
 
 public class EntryValutazione {
