+++ src/main/java/com/example/bookrecommenderdev/client/model/Libro.java
@@ -1,4 +1,4 @@
-package com.example.bookrecommenderdev.model;
+package com.example.bookrecommenderdev.client.model;
 
 import java.util.UUID;
